fix: update windows panel detail info
