+++ src/renderer/StatusBoard.tsx
@@ -9,6 +9,7 @@ import { EVENT_DATA_CHANNEL_NAME, EVENT_ACTION_CHANNEL_NAME } from '../common/co
 
 import styles from './StatusBoard.module.less';
 import type { ProcessInfo } from '../common/interface';
+import { findCommonStringPlus, findName, isWindows } from './util';
 
 interface MemoryStats {
   workingSetSize: number;
@@ -18,64 +19,6 @@ interface MemoryStats {
 interface CpuStatus {
   percentCPUUsage: number;
 }
-
-// 提取双引号的内容
-const PROCESS_REGEX = /"?([^"]*)"?/;
-
-/**
- * 获取进程 cmd 的简易信息
- */
-const findName = (cmd?: string, startIndex = 0): string => {
-  if (!cmd) {
-    return '';
-  }
-  const index = cmd.indexOf(' -');
-  const end = index !== -1 ? index : cmd.length;
-  if (startIndex > 0) {
-    return `...${cmd.substring(startIndex + 1, end)}`;
-  }
-  const simplePath = cmd.substring(0, end);
-  const matched = simplePath.match(PROCESS_REGEX);
-  if (matched) {
-    return matched[1];
-  }
-  return simplePath;
-};
-
-/**
- * 找出相同的字符串
- */
-const findCommonString = (str1?: string, str2?: string) => {
-  let index = -1;
-  if (typeof str1 !== 'string' || typeof str2 !== 'string') {
-    return index;
-  }
-
-  const minLen = Math.min(str1.length, str2.length);
-  for (let i = 0; i < minLen; ++i) {
-    if (str1[i] === str2[i]) {
-      index = i;
-    } else {
-      break;
-    }
-  }
-  return index;
-};
-
-/**
- * 从多个字符串中找到相同字符串
- */
-const findCommonStringPlus = (strs: string[]) => {
-  if (strs.length < 2) {
-    return -1;
-  }
-
-  const indexes = [];
-  for (let i = 1; i < strs.length; i++) {
-    indexes.push(findCommonString(strs[i - 1], strs[i]));
-  }
-  return Math.min(...indexes);
-};
 
 // 测试共同字符串数量
 const MAX_COMMON_STRING_TEST = 3;
@@ -170,7 +113,7 @@ const useViewModel = (props: StatusBoardProps) => {
 
   const updateAppMetrics = (_: any, appMetrics: any[]) => {
     const list = Array.isArray(appMetrics) ? appMetrics : [];
-    if (list.length > MAX_COMMON_STRING_TEST) {
+    if ((list.length > MAX_COMMON_STRING_TEST) && !isWindows()) {
       setProcessBaseIndex(findCommonStringPlus(list.map((item) => item.cmd)));
     }
