+++ src/renderer/components/ProcessDetail/index.module.less
@@ -1,7 +1,7 @@
 .wrapper {
   display: flex;
   flex-direction: column;
-  height: 300px;
+  height: 150px;
 
   .detail {
     margin: 8px;
