+++ src/renderer/components/WebContentsDetail/index.module.less
@@ -1,7 +1,7 @@
 .wrapper {
   display: flex;
   flex-direction: column;
-  height: 300px;
+  height: 150px;
 
   .info {
     margin-left: 8px;
@@ -9,6 +9,7 @@
 
   .detail {
     margin: 8px;
+    flex: 1;
     font-size: 12px;
     flex: 1;
     overflow-wrap: break-word;
